requirements.txt
boto3==1.13.11
botocore==1.16.11
dj-database-url==0.5.0
Django==2.1
django-cors-headers==3.2.0
django-storages==1.9.1
djangorestframework==3.10.3
docutils==0.15.2
gunicorn==20.0.0
jmespath==0.10.0
Pillow==6.2.1
psycopg2-binary==2.8.4
python-dateutil==2.8.1
pytz==2019.3
s3transfer==0.3.3
six==1.14.0
urllib3==1.25.9
whitenoise==4.1.4
